mb/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mainbranch"
version = "0.1.0.dev0"
description = "Main Branch engine umbrella — scaffolds, validates, and graphs business-as-files repos. Built for Claude Code."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Devon Meadows", email = "devon@noontide.co" }
]
keywords = [
  "cli",
  "business-as-files",
  "claude-code",
  "skills",
  "scaffolding",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Utilities",
]
dependencies = [
  "typer>=0.12",
  "pyyaml>=6.0",
  "rich>=13.0",
]

[project.optional-dependencies]
ogrender = [
  "cairosvg>=2.7",
]
dev = [
  "ruff~=0.6",
  "mypy~=1.10",
  "pytest~=8.0",
  "pytest-cov~=5.0",
  "types-PyYAML",
]

[project.scripts]
mb = "mb.cli:app"

[project.urls]
Homepage = "https://github.com/mainbranch-ai/mb-vip"
Repository = "https://github.com/mainbranch-ai/mb-vip"
Issues = "https://github.com/mainbranch-ai/mb-vip/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mb*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
mb = ["_data/**/*", "py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["yaml.*", "rich.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
